requirements.txt
Flask==2.3.3
Pillow==10.0.0
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.2
importlib-metadata==6.8.0
SQLAlchemy==2.0.20
flask-session==0.5.0
pytest==7.4.0
python-dotenv==1.0.0 